+++ backend/canisters/group/impl/src/timer_job_types.rs
@@ -9,9 +9,8 @@ use constants::{DAY_IN_MS, MINUTE_IN_MS, NANOS_PER_MILLISECOND, SECOND_IN_MS};
 use ledger_utils::process_transaction;
 use serde::{Deserialize, Serialize};
 use tracing::error;
-use types::{
-    BlobReference, CanisterId, MessageId, MessageIndex, P2PSwapStatus, PendingCryptoTransaction, TimestampMillis, UserId,
-};
+use types::icrc1::Account;
+use types::{BlobReference, MessageId, MessageIndex, P2PSwapStatus, PendingCryptoTransaction, TimestampMillis, UserId};
 
 #[derive(Serialize, Deserialize, Clone)]
 pub enum TimerJob {
@@ -266,11 +265,11 @@ impl Job for FinalPrizePaymentsJob {
 
 impl Job for MakeTransferJob {
     fn execute(self) {
-        let sender = read_state(|state| state.env.canister_id());
+        let sender = read_state(|state| Account::from(state.env.canister_id()));
         let pending = self.pending_transaction.clone();
         ic_cdk::futures::spawn_migratory(make_transfer(pending, sender, self.attempt));
 
-        async fn make_transfer(mut pending_transaction: PendingCryptoTransaction, sender: CanisterId, attempt: u32) {
+        async fn make_transfer(mut pending_transaction: PendingCryptoTransaction, sender: Account, attempt: u32) {
             if let Err(error) = process_transaction(pending_transaction.clone(), sender, true).await {
                 error!(?error, "Transaction failed");
                 if attempt < 50 {